% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{Calculate Minimum Sample Size for T-Tests}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{A numeric vector of preliminary data for the first sample.}

\item{x2}{An optional numeric vector of preliminary data for the second sample.}
}
\value{
The minimum sample size needed for each group to achieve the desired power.
}
\description{
This function is a wrapper around pwr::pwr.t2n.test that takes either one ( x1 )
or two ( x2 ) samples of preliminary data and returns the minimum sample size needed
for a t-test of the null hypotheses that either μ_x1==0 or μ_x1==μ_x2 with 80% power at alpha=0.05.
Note: if a very small sample size is required as the effect size is very large and pwr calculation
fails, we will just return 2 (This is also the minimum sample size set in this function).
}
\examples{
x1 = c(0.5, 0.1, -0.2, 0.3, 0.1, -0.1, 0, 0.2, 0.3, -0.4)
x2 = c(5, 5, 5, 5, 5)
minimumN(x1)#One-sample
minimumN(x1, x2)#Two-sample
}
