% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize Column Names to camelCase}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A tibble whose column names you want to standardize.}
}
\value{
A tibble with standardized column names in camelCase.
}
\description{
This function is a wrapper around dplyr::rename_with and janitor::make_clean_names that
converts the variables in a tibble data to "small_camel" case.
}
\examples{
#Creating a sample tibble
sample_data = tibble(`first column` = 1:5, `second column` = 6:10)
#Standardizing names
new_data = standardizeNames(sample_data)
print(new_data)
}
