% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCurv.R
\name{survCurv}
\alias{survCurv}
\title{Plot the survival curve using only tidyverse}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{A numerical vector indicating the status at time}

\item{time}{A numerical vector indicating the time at corresponding status}
}
\value{
A plot of the survival curve S(t).
}
\description{
This function takes numerical vectors status and a numerical vector time, and calculates
and plots a survival curve S(t).
}
\examples{
data = read.csv("https://jlucasmckay.bmi.emory.edu/global/bmi510/Labs-Materials/survival.csv")
survCurv(data$status, data$time)
}
