% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Unscale vectors that has been passed through scale()}
\usage{
unscale(x)
}
\arguments{
\item{x}{A numeric vector that has been scaled.}
}
\value{
A numeric vector with the scaling reversed.
}
\description{
This function takes a vector that has been put through scale
and reverses the centering/scaling, if any.
}
\examples{
original_vector = c(1, 2, 3, 4, 5)
scaled_vector = scale(original_vector)
unscaled_vector = unscale(scaled_vector)
}
